requirements.txt
hausdorff
h5py
einops
matplotlib
numba
numpy
opencv_python
pandas
Pillow
seaborn
setuptools
scikit-image
tensorboard
tqdm
thop